python/packages/autogen-kubernetes-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autogen-kubernetes-mcp"
version = "0.5.1"
description = "MPC server provides code interpreter with kubernetes workload"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "autogen-kubernetes>=0.5.1",
    "mcp>=1.12.2",
]

[project.scripts]
autogen-kubernetes-mcp = "autogen_kubernetes_mcp.server:main"

[tool.hatch.build.target.wheel]
packages = ["src/autogen_kubernetes_mcp"]

[tool.ruff]
extend = "../../pyproject.toml"
include = ["src/**", "tests/*.py"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "module"

[tool.poe]
include = "../../shared_tasks.toml"

[tool.poe.tasks]
test = "pytest -n auto"

[tool.mypy]
[[tool.mypy.overrides]]
module = "kubernetes.*"
ignore_missing_imports = true